[core] Correct logic in battleutils.cpp uint8...
...to properly display higher level bp for Carbuncle

* Winter found this method to be satisfactory and resolved the issue of the bp not displaying correctly
* Based on retail captures

+++ src/map/utils/battleutils.cpp
@@ -88,7 +88,7 @@ std::array<std::array<uint16, MAX_SKILLCHAIN_COUNT + 1>, MAX_SKILLCHAIN_LEVEL + 
 
 std::array<CWeaponSkill*, MAX_WEAPONSKILL_ID> g_PWeaponSkillList; // Holds all Weapon skills
 std::array<CMobSkill*, MAX_MOBSKILL_ID>       g_PMobSkillList;    // List of mob skills
-std::unordered_map<uint8, CPetSkill*>         g_PPetSkillList;    // List of pet skills
+std::unordered_map<uint32, CPetSkill*>         g_PPetSkillList;    // List of pet skills
 
 std::array<std::list<CWeaponSkill*>, MAX_SKILLTYPE> g_PWeaponSkillsList;
 std::unordered_map<uint16, std::vector<uint16>>     g_PMobSkillLists; // List of mob skills defined from mob_skill_lists.sql
@@ -490,6 +490,7 @@ namespace battleutils
 
     CPetSkill* GetPetSkill(uint16 SkillID)
     {
+
         if (g_PPetSkillList.find(SkillID) != g_PPetSkillList.end())
         {
             return g_PPetSkillList[SkillID];
